Move MenuMatcher into .swtbot, and rewrite WithRegex expression
