+++ plugins/com.google.cloud.tools.eclipse.integration.appengine/build.properties
@@ -5,5 +5,4 @@ bin.includes = META-INF/,\
                test-projects/,\
                plugin.properties
 additional.bundles = org.eclipse.equinox.common,\
-                     org.eclipse.ui.console,\
-                     org.apache.log4j
+                     org.eclipse.ui.console
