+++ plugins/com.google.cloud.tools.eclipse.swtbot/META-INF/MANIFEST.MF
@@ -20,4 +20,5 @@ Export-Package: com.google.cloud.tools.eclipse.swtbot;
    org.eclipse.swtbot.eclipse.finder.widgets,
    org.eclipse.swtbot.swt.finder"
 Import-Package: com.google.common.base;version="[21.0.0,22.0.0)",
- org.eclipse.wst.validation
+ org.eclipse.wst.validation,
+ org.hamcrest;version="1.3.0"
